add tilemap #2 & add camera

+++ src/Engine/Graphics/TilemapOrientationMatrix.cs
@@ -4,12 +4,12 @@ namespace Scorpia.Engine.Graphics;
 
 internal struct TilemapOrientationMatrix
 {
-    public readonly double f0, f1, f2, f3;
-    public readonly double b0, b1, b2, b3;
-    public readonly double startAngle;
+    public readonly float f0, f1, f2, f3;
+    public readonly float b0, b1, b2, b3;
+    public readonly float startAngle;
 
-    private TilemapOrientationMatrix(double f0, double f1, double f2, double f3, double b0, double b1, double b2, double b3,
-        double startAngle)
+    private TilemapOrientationMatrix(float f0, float f1, float f2, float f3, float b0, float b1, float b2, float b3,
+        float startAngle)
     {
         this.f0 = f0;
         this.f1 = f1;
@@ -22,15 +22,15 @@ internal struct TilemapOrientationMatrix
         this.startAngle = startAngle;
     }
 
-    public static TilemapOrientationMatrix Pointy { get; } = new(Math.Sqrt(3.0), Math.Sqrt(3.0) / 2.0, 0.0,
-        3.0 / 2.0,
-        Math.Sqrt(3.0) / 3.0, -1.0 / 3.0, 0.0, 2.0 / 3.0,
-        0.5);
+    public static TilemapOrientationMatrix Pointy { get; } = new((float)Math.Sqrt(3.0), (float)Math.Sqrt(3.0) / 2.0f, 0.0f,
+        3.0f / 2.0f,
+        ((float)Math.Sqrt(3.0)) / 3.0f, -1.0f / 3.0f, 0.0f, 2.0f / 3.0f,
+        0.5f);
 
-    public static TilemapOrientationMatrix Flat { get; } = new(3.0 / 2.0, 0.0, Math.Sqrt(3.0) / 2.0,
-        Math.Sqrt(3.0),
-        2.0 / 3.0, 0.0, -1.0 / 3.0, Math.Sqrt(3.0) / 3.0,
-        0.0);
+    public static TilemapOrientationMatrix Flat { get; } = new(3.0f / 2.0f, 0.0f, (float)Math.Sqrt(3.0) / 2.0f,
+        (float)Math.Sqrt(3.0),
+        2.0f / 3.0f, 0.0f, -1.0f / 3.0f, (float)Math.Sqrt(3.0) / 3.0f,
+        0.0f);
 }
 
 public enum TilemapOrientation
